requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
DictObject==1.1.1
filelock==3.0.12
huggingface-hub==0.0.12
idna==3.2
joblib==1.0.1
luckydonald-utils==0.83
numpy==1.21.1
packaging==21.0
pyparsing==2.4.7
pytg==0.4.10
python-telegram-bot==13.7
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
sacremoses==0.0.45
six==1.16.0
tokenizers==0.10.3
torch==1.9.0
tornado==6.1
tqdm==4.62.0
transformers==4.9.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
